pyproject.toml
[project]
name = "strands-redis-session-manager"
version = "0.1.0"
description = "Redis-based session management for persisting and managing agent sessions in the Strands Agents SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "redis>=7.0.1",
    "strands-agents>=1.16.0",
]
license = "MIT"
authors = [{ name = "lekhnath", email = "inboxlekhnath@gmail.com" }]
keywords = ["strands", "agents", "session", "redis", "ai", "llm", "memory"]

[project.urls]
Homepage = "https://github.com/lekhnath/strands-redis-session-manager"
Documentation = "https://github.com/lekhnath/strands-redis-session-manager#readme"
Repository = "https://github.com/lekhnath/strands-redis-session-manager"
Issues = "https://github.com/lekhnath/strands-redis-session-manager/issues"

[dependency-groups]
dev = [
    "fakeredis>=2.24.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.5",
]
